Functionality complete

+++ pillowtalk-frontend/src/Components/DreamCard/DreamCard.js
@@ -1,16 +1,44 @@
-import React from "react";
+import React, { useState } from "react";
 import "./DreamCard.style.css"
 
-function DreamCard({ dream, handleOnDelete }){
+
+function DreamCard({ dream, handleOnDelete, setIsEditing, currentUser }){
     const { id, name, date, description, image_url, mood, rating } = dream;
 
     function handleDeleteClick(){
       console.log(id);
-      fetch(`https://damp-beach-45746.herokuapp.com/dreams/${id}`, {
+      if (dream.user_id === currentUser.id) {
+        fetch(`https://damp-beach-45746.herokuapp.com/dreams/${id}`, {
         method: "DELETE"
       })
       handleOnDelete(id)
+      }
+      else {
+        alert("Permission denied to delete")
+      }
     }
+
+    const [editDream, setEditDream]= useState({
+      name: "",
+      date: "",
+      description:"",
+      rating: 0,
+      mood: "",
+      image_url: "",
+      user_id: 1
+     })
+
+     function handleClick(){
+       console.log(dream)
+       console.log(currentUser)
+      if (dream.user_id === currentUser.id) {
+        setIsEditing(true)
+      }
+      else {
+       alert("Permission denied to edit");
+      }
+    }
+ 
 
     return(
       <div>
@@ -23,6 +51,7 @@ function DreamCard({ dream, handleOnDelete }){
           <img id ="dreamImgs" src={image_url} />
         </li>
         <button onClick={handleDeleteClick}> Delete </button>
+        <button onClick={handleClick} > Edit Dream </button>
       </div>
     )
 }
